refactor screens
and its imports
