+++ src/screens/Article/GoalTopBubble/TopBubble/TopBubble.tsx
