+++ src/screens/ChattingRoom/ChatsContainer/ChatBox/SystemMessage/SystemMessage.tsx
