+++ src/screens/Profile/Profile.tsx
@@ -3,8 +3,8 @@ import { useSelector } from 'react-redux';
 
 import { RootState } from '@/store';
 
-import LoggedProfile from './Logged';
-import UnloggedProfile from './UnLogged';
+import LoggedProfile from './Logged/Logged';
+import UnloggedProfile from './UnLogged/UnLogged';
 
 function Profile(): JSX.Element {
   const logged = useSelector((state: RootState) => state.user.logged);
