+++ src/screens/Search/SearchedArticle/SearchedList/SearchedList.tsx
