+++ src/screens/index.ts
@@ -1,19 +1,19 @@
-export { default as AppLoading } from './AppLoading';
-export { default as Article } from './Article';
-export { default as ChattingRoom } from './ChattingRoom';
-export { default as Search } from './Search';
-export { default as ChattingList } from './ChattingList';
-export { default as Drawer } from './Drawer';
-export { default as Home } from './Home';
-export { default as Login } from './Login';
-export { default as Notification } from './Notification';
-export { default as Profile } from './Profile';
-export { default as ProfileModify } from './ProfileModify';
-export { default as SignUp } from './SignUp';
-export { default as TOS } from './TOS';
-export { default as WriteArticle } from './WriteArticle';
-export { default as EditArticle } from './EditArticle';
+export { default as AppLoading } from './AppLoading/AppLoading';
+export { default as Article } from './Article/Article';
+export { default as ChattingRoom } from './ChattingRoom/ChattingRoom';
+export { default as Search } from './Search/Search';
+export { default as ChattingList } from './ChattingList/ChattingList';
+export { default as Drawer } from './Drawer/Drawer';
+export { default as Home } from './Home/Home';
+export { default as Login } from './Login/Login';
+export { default as Notification } from './Notification/Notification';
+export { default as Profile } from './Profile/Profile';
+export { default as ProfileModify } from './ProfileModify/ProfileModify';
+export { default as SignUp } from './SignUp/SignUp';
+export { default as TOS } from './TOS/TOS';
+export { default as WriteArticle } from './WriteArticle/WriteArticle';
+export { default as EditArticle } from './EditArticle/EditArticle';
 
 // TODO: ssu1018
 //  @woohm402: 이거 왜 여기서 export하죠..?
-export { default as SearchedArticle } from './Search/SearchedArticle';
+export { default as SearchedArticle } from './Search/SearchedArticle/SearchedArticle';
